+++ src/DotNext.Tests/Runtime/InteropServices/UnmanagedFunctionTests.cs
@@ -2,7 +2,6 @@
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.InteropServices;
-using Xunit;
 
 namespace DotNext.Runtime.InteropServices
 {