+++ src/DotNext.Unsafe/IO/MemoryMappedFiles/MemoryMappedViewAccessorExtensions.cs
@@ -1,0 +1,17 @@
+﻿using System.IO;
+using System.IO.MemoryMappedFiles;
+
+namespace DotNext.IO.MemoryMappedFiles
+{
+    internal static class MemoryMappedViewAccessorExtensions
+    {
+        internal static FileAccess GetFileAccess(this MemoryMappedViewAccessor accessor)
+            => (accessor.CanRead.ToInt32() + (accessor.CanWrite.ToInt32() << 1)) switch
+            {
+                1 => FileAccess.Read,
+                2 => FileAccess.Write,
+                3 => FileAccess.ReadWrite,
+                _ => default,
+            };
+    }
+}