+++ src/DotNext/StringExtensions.cs
@@ -3,7 +3,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 using static InlineIL.IL;
 using static InlineIL.IL.Emit;
-using M = InlineIL.MethodRef;
 using Var = InlineIL.LocalVar;
 
 namespace DotNext
@@ -91,11 +90,26 @@ namespace DotNext
             Conv_U();
             Dup();
             Brfalse(methodExit);
-            Call(M.PropertyGet(typeof(RuntimeHelpers), nameof(RuntimeHelpers.OffsetToStringData)));
+            Push(RuntimeHelpers.OffsetToStringData);
             Conv_U();
             Add();
             MarkLabel(methodExit);
             return ref ReturnRef<char>();
+        }
+
+        /// <summary>
+        /// Extracts substring from the given string.
+        /// </summary>
+        /// <remarks>
+        /// This method if useful for .NET languages without syntactic support of ranges.
+        /// </remarks>
+        /// <param name="str">The instance of string.</param>
+        /// <param name="range">The range of substring.</param>
+        /// <returns>The part of <paramref name="str"/> extracted according with supplied range.</returns>
+        public static string Substring(this string str, Range range)
+        {
+            var (start, length) = range.GetOffsetAndLength(str.Length);
+            return str.Substring(start, length);
         }
     }
 }